requirements.txt
mysql-connector-python==9.3.0
pandas==2.3.0
matplotlib==3.10.3